2010 CAGONT Poster for Development of Maps for Ryerson Alumni Relations
This is the digital version (.pdf, 10MB) of the poster I submitted to the Canadian Association of Geographers (Ontario Division) conference in 2010. The poster illustrates the process I used to complete work for Ryerson University's Alumni Relations department. They required maps of where Ryerson Alumni resided in 2010 and these maps present this information. The maps appeared in their annual report.